评论接口返回加用户昵称，头像

+++ reading-website-api/src/main/java/com/reading/website/api/domain/UserBaseInfoDO.java
@@ -19,12 +19,12 @@ public class UserBaseInfoDO implements Serializable {
     /**
      * 主键userId
      */
-    private Long id;
+    private Integer id;
 
     /**
      * userId List
      */
-    private List<Long> ids;
+    private List<Integer> ids;
 
     /**
      * 创建时间
